Add files via upload
This component is almost specifically in charge of tracking the NFT boost attributes, which are as follows:
- 1 NFT held will boost rewards for 5 Nodes;
- the life span of the nodes will decrease by the same percentage of boosted rewards relative to time. 
- percentages are modular, this particular contract is set up with 3 tiers; allowing 25%, 50% 0r 75% boost mechanisms.

all of the attributes are delegated to the nodes via first in first out with as much dedication put into time tracking as possible. The mechanism itself relies on an external NFT staking contract in which this boost contract and the ecosystems oracleized contract aid in co-calculations.