Add files via upload
This contract became largely redundant as development went on as the fees and boost seemed to be the most defining attributes, however it was kept for traditional purposes as well it adding to the overall struct storage. Its largest attribute to note is that it keeps the most precise records of collapsed and running nodes as well as overflow record keeping for odd ends such as boost records.