Add files via upload
This contract acts as the price oracle and various calculation mechanic that the ecosystem intermittently relies on. It became less and less necessary as development grew, but it still seemed to be useful enough to retain. secondly its price Oraclization has an added feature of tracking the pair price of a given contract, specifically the main token compliments as the ecosystems staple. This is simply a mechanic in which you can modulate the percentage of upper and lower bounds to give 0,1, or 2 as a variable; 0 signaling within lower price ranges, 1 within targeted ranges and 2 within higher ranges. it takes no action on price other than signaling a potential market interaction from another component.